Added dialogue (could shift this to the wiki)
Dear Mr Glixan,
I I'm very sorry to bring you the sad news of the Shai Hulud.

As you know she has been with us for many years serving up randomised divinity selections in a discordant manner to the many surfaces and paradise seekers roaming the web looking for answers to their questions.

as you may also be aware shai hulud has found a permanent or what seemed to be a permanent refuge in the the metadata of synapse New Zealand an organisation dedicated to protect the preserve of Shahid Masood.

alas when I went to visit the website a few days ago I discovered that shy Helens flash meta language had finally become redundant... The actionscript had become one with the universe.. in short shai hulud has 'parsed' away from its html moorings and into the greater surf..

I tried the obvious and most immediate solutions including parse a html file in python and bypassing the action strip. But Aliah! I have found a new solution on which we can collaborate it is based on github... and a bastardization of the JavaScript divinity randomised answer selector of the still running and finally surviving oracle of professor brainstorm!

You will receive a link in a forth coming email with which we can replace the brain of Dr brainstorm with the answers 2020 version of Che halud.. The Cuban revolutionary version of the divinity select oracle Twenty20 shall be permanently resident on the synapse of New Zealand org


Salutations,
Guardians of the Shai Halud oracle, cyberspace, electrom.
Dr Amit NagaRaja